api/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
redis==5.0.1
pyjwt==2.8.0
azure-communication-email==1.0.0
python-multipart==0.0.6 